Rename Hungman_Poject.py to Hungman_Poject
The hangman game is a simple game. In this game, player selects a word and then other players have a certain number of guesses to guess the characters in the word. If the players are able to guess the characters in the entire word within certain attempts, they win. Otherwise, they lose.